Introduce new `User` concept
Client broadly encompassed three different concerns:

* User with all its metadata
* Session information
* Connection handles

While session and connection information and connection information are
reliant on each other, the concept of a user that just exists in the
system is not.

A user can be registered and removed at any point in time and does not
need to start a session at the same time. Additionally, a user can
have multiple sessions with differing connection handles so it doesn't
make sense to conflate these concepts anymore.

In this commit, we reduce `Clients` functionality, preparing `Client` to
become the `Session`.